+++ R/FLOW.R
@@ -31,6 +31,7 @@ library(torch)
 #' }
 #' @examples
 #' \donttest{
+#'if (torch_available()) {
 #'flow <- normalizing_flow(c(2,5,5), transform_type='RNVP', num_transforms = 3)
 #'flow$to(device = 'cpu')
 #'x <- torch::torch_rand(2, device = 'cpu')
@@ -38,7 +39,7 @@ library(torch)
 #'z_out <- output$z
 #'print(dim(z_out))
 #'log_det <- output$logdet
-#'print(log_det)
+#'print(log_det)}
 #' }
 #' @export
 normalizing_flow <- torch::nn_module(