+++ R/RNVP.R
@@ -57,11 +57,12 @@ mlp <- torch::nn_module(
 #' }
 #' @examples
 #' \donttest{
+#' if (torch_available()) {
 #' z <- torch::torch_rand(200)
 #' layer <- rnvp_layer(c(200,50,100))
 #' out <- layer(z)
 #' print(dim(out))
-#' print(layer$log_det())
+#' print(layer$log_det())}
 #' }
 #' @export
 rnvp_layer <- torch::nn_module(