+++ tests/testthat/test_smoke_training.R
@@ -1,12 +1,10 @@
-test_that("Smoke: tiny model trains one epoch", {
+test_that("Smoke: tiny model trained for one epoch", {
   testthat::skip_on_cran()
   if (!requireNamespace("torch", quietly = TRUE)) {
     testthat::skip("torch not available")
   }
-
   i <- 5000
   j <- 15
-
   set.seed(2)
   torch::torch_manual_seed(2)
   #generate some data